Integer-size warning related fix
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@12659 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/sphinxbase/bio.h
@@ -294,7 +294,7 @@ int16* bio_read_wavfile(char const *directory, /**< In: the folder where the fil
 			char const *extension, /**< In: file extension */
 			int32 header,          /**< In: the size of the header to skip usually 44 bytes */
 			int32 endian,          /**< In: endian of the data */
-			int32 *nsamps          /**< Out: number of samples read */
+			size_t *nsamps         /**< Out: number of samples read */
 			);
 
 #ifdef __cplusplus
